Create Exercise-1-ArrayDemo
Write a program named ArrayDemo that stores an array of 10 integers. Until the user
enters a sentinel value, allow the user four options: (1) to view the list in order from
the first to last position in the stored array, (2) to view the list in order from the last to
first position, (3) to choose a specific position to view, or (4) to quit the application.